+++ src/peplum/app/data/config.py
@@ -3,7 +3,7 @@
 ##############################################################################
 # Python imports.
 from contextlib import contextmanager
-from dataclasses import asdict, dataclass
+from dataclasses import asdict, dataclass, field
 from functools import lru_cache
 from json import dumps, loads
 from pathlib import Path
@@ -43,6 +43,9 @@ class Configuration:
 
     peps_sort_reversed: bool = False
     """Should the PEPs sort order be reversed?"""
+
+    bindings: dict[str, str] = field(default_factory=dict)
+    """Command keyboard binding overrides."""
 
 
 ##############################################################################