+++ ChangeLog.md
@@ -1,5 +1,12 @@
 # Peplum ChangeLog
 
+## Unreleased
+
+**Released: WiP**
+
+- Added the ability to configure the keyboard bindings for the system
+  commands.
+
 ## v0.5.0
 
 **Released: 2025-04-08**